refactor(LottoController): 컨트롤러 리팩토링
- 4개의 컨트롤러가 각각의 역할을 가지도록 함
- LottoController가 핵심 흐름만 가지도록 수정
